requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.1.1
typing_extensions==4.14.1
yarl==1.20.1
